rompe todo en el gamelop

+++ src/Client/Controllers/GameController.h
@@ -1,27 +1,28 @@
 #ifndef GAMECONTROLLER_H
 #define GAMECONTROLLER_H
 
+#include "../../Common/DTO/dto.h"
+#include "../../Common/Types/command.h"
 #include "../Protocol/deserializer.h"
 #include "../Protocol/serializer.h"
-#include "../../Common/Types/command.h"
-#include "../../Common/DTO/dto.h"
 
 class GameController {
 private:
     Serializer& serializer;
     Deserializer& deserializer;
     std::shared_ptr<Queue<std::unique_ptr<DTO>>>& gameQueue;
+
 public:
-    GameController(Serializer& serializer, Deserializer& deserializer, std::shared_ptr<Queue<std::unique_ptr<DTO>>>& gameQueue);
+    GameController(Serializer& serializer, Deserializer& deserializer,
+                   std::shared_ptr<Queue<std::unique_ptr<DTO>>>& gameQueue);
 
-    void sendMsg(uint32_t playerId, Command& cmd, std::vector<uint8_t>& parameters);
+    void sendMsg(uint8_t playerId, Command& cmd, std::vector<uint8_t>& parameters);
 
-    void move_msg(uint32_t playerId, std::vector<uint8_t>& parameters);
+    void move_msg(uint8_t playerId, std::vector<uint8_t>& parameters);
 
-    void shoot_msg(uint32_t playerId);
+    void shoot_msg(uint8_t playerId);
 
     std::unique_ptr<DTO> getServerMsg();
-	
 };
 
 #endif  // GAMECONTROLLER_H
