+++ src/Common/DTO/gamesList.cpp
@@ -5,7 +5,7 @@
 
 GamesListDTO::GamesListDTO(): CommandDTO(Command::GAMES_LIST) {}
 
-GamesListDTO::GamesListDTO(std::map<uint32_t, GameInfo> games):
+GamesListDTO::GamesListDTO(std::map<uint8_t, GameInfo> games):
         CommandDTO(Command::GAMES_LIST), games(std::move(games)) {}
 
-std::map<uint32_t, GameInfo> GamesListDTO::getGames() const { return games; }
+std::map<uint8_t, GameInfo> GamesListDTO::getGames() const { return games; }
