+++ src/Common/DTO/switchWeapon.cpp
@@ -1,5 +1,5 @@
 #include "switchWeapon.h"
 
-SwitchWeaponDTO::SwitchWeaponDTO(uint32_t& playerId, WeaponType type): type(type) {}
+SwitchWeaponDTO::SwitchWeaponDTO(uint8_t& playerId, WeaponType type): type(type) {}
 
 WeaponType SwitchWeaponDTO::getWeaponType() const { return type; }
