+++ src/Common/Types/lobbyMessage.h
@@ -13,9 +13,9 @@ struct LobbyMessage {
     Command cmd;
     int8_t maxPlayers;
     std::string gameName;
-    uint32_t gameId;
+    uint8_t gameId;
     CharacterType character;
-    uint32_t mapId;
+    uint8_t mapId;
 
     LobbyMessage():
             playerName(""),
@@ -27,8 +27,8 @@ struct LobbyMessage {
             mapId(-1) {}
 
     LobbyMessage(std::string playerName, const Command& cmd, const int8_t& maxPlayers,
-                 std::string gameName, const uint32_t& gameId, const CharacterType& character,
-                 const uint32_t& mapId):
+                 std::string gameName, const uint8_t& gameId, const CharacterType& character,
+                 const uint8_t& mapId):
             playerName(std::move(playerName)),
             cmd(cmd),
             maxPlayers(maxPlayers),
@@ -45,11 +45,11 @@ struct LobbyMessage {
 
     void setGameName(const std::string& name) { gameName = name; }
 
-    void setGameId(const uint32_t& id) { gameId = id; }
+    void setGameId(const uint8_t& id) { gameId = id; }
 
     void setCharacter(CharacterType selectedCharacter) { character = selectedCharacter; }
 
-    void setMap(const uint32_t& id) { mapId = id; }
+    void setMap(const uint8_t& id) { mapId = id; }
 
     std::string getPlayerName() const { return playerName; }
 
@@ -59,15 +59,15 @@ struct LobbyMessage {
 
     std::string getGameName() const { return gameName; }
 
-    uint32_t getGameId() const { return gameId; }
+    uint8_t getGameId() const { return gameId; }
 
     CharacterType getCharacter() const { return character; }
 
-    uint32_t getMap() const { return mapId; }
+    uint8_t getMap() const { return mapId; }
 
     LobbyMessage& build(const std::string& namePlayer, Command lobbyCmd, int8_t maxPlayerNumber,
-                        const std::string& nameGame, uint32_t idGame, CharacterType characterType,
-                        const uint32_t& idMap) {
+                        const std::string& nameGame, uint8_t idGame, CharacterType characterType,
+                        const uint8_t& idMap) {
         this->playerName = namePlayer;
         this->cmd = lobbyCmd;
         this->maxPlayers = maxPlayerNumber;
