+++ src/Common/queueMonitor.cpp
@@ -8,7 +8,7 @@
 #include "DTO/game.h"
 
 template <typename T>
-std::shared_ptr<Queue<T>> QueueMonitor<T>::createQueue(uint32_t id) {
+std::shared_ptr<Queue<T>> QueueMonitor<T>::createQueue(uint8_t id) {
     std::lock_guard<std::mutex> lock(mtx);
     auto queue = std::make_shared<Queue<T>>();
     queues[id].push_back(queue);
@@ -16,7 +16,7 @@ std::shared_ptr<Queue<T>> QueueMonitor<T>::createQueue(uint32_t id) {
 }
 
 template <typename T>
-void QueueMonitor<T>::closeQueue(uint32_t id, std::shared_ptr<Queue<T>> queue) {
+void QueueMonitor<T>::closeQueue(uint8_t id, std::shared_ptr<Queue<T>> queue) {
     std::lock_guard<std::mutex> lock(mtx);
     auto& vec = queues[id];
     auto it = std::find(vec.begin(), vec.end(), queue);
@@ -26,7 +26,7 @@ void QueueMonitor<T>::closeQueue(uint32_t id, std::shared_ptr<Queue<T>> queue) {
 }
 
 template <typename T>
-void QueueMonitor<T>::removeQueue(uint32_t id, std::shared_ptr<Queue<T>> queue) {
+void QueueMonitor<T>::removeQueue(uint8_t id, std::shared_ptr<Queue<T>> queue) {
     std::lock_guard<std::mutex> lock(mtx);
     auto& vec = queues[id];
     auto it = std::find(vec.begin(), vec.end(), queue);
@@ -39,7 +39,7 @@ void QueueMonitor<T>::removeQueue(uint32_t id, std::shared_ptr<Queue<T>> queue) 
 }
 
 template <typename T>
-void QueueMonitor<T>::closeQueues(uint32_t id) {
+void QueueMonitor<T>::closeQueues(uint8_t id) {
     std::lock_guard<std::mutex> lock(mtx);
     auto& vec = queues[id];
     for (auto& queue: vec) {
@@ -48,13 +48,13 @@ void QueueMonitor<T>::closeQueues(uint32_t id) {
 }
 
 template <typename T>
-void QueueMonitor<T>::removeQueues(uint32_t id) {
+void QueueMonitor<T>::removeQueues(uint8_t id) {
     std::lock_guard<std::mutex> lock(mtx);
     queues.erase(id);
 }
 
 template <typename T>
-void QueueMonitor<T>::broadcast(uint32_t id, T&& event) {
+void QueueMonitor<T>::broadcast(uint8_t id, T&& event) {
     std::lock_guard<std::mutex> lock(mtx);
     auto it = queues.find(id);
     if (it != queues.end()) {
@@ -73,7 +73,7 @@ void QueueMonitor<T>::broadcast(uint32_t id, T&& event) {
 
 
 template <typename T>
-void QueueMonitor<T>::assignGameIdToQueues(uint32_t gameId, std::shared_ptr<Queue<T>> queue) {
+void QueueMonitor<T>::assignGameIdToQueues(uint8_t gameId, std::shared_ptr<Queue<T>> queue) {
     std::lock_guard<std::mutex> lock(mtx);
     auto& vec = queues[gameId];
     if (!queue) {
