+++ src/Server/CommandHandlers/Lobby/startGame.cpp
@@ -16,8 +16,8 @@ StartGameHandler::StartGameHandler(std::unique_ptr<StartGameDTO> command):
 void StartGameHandler::execute(GameMonitor& gameMonitor, std::atomic<bool>& inGame,
                                std::shared_ptr<Queue<std::unique_ptr<CommandDTO>>> recvQueue,
                                std::shared_ptr<Queue<std::unique_ptr<DTO>>> sendQueue) {
-    uint32_t playerId = command->getPlayerId();
-    uint32_t gameId = command->getGameId();
+    uint8_t playerId = command->getPlayerId();
+    uint8_t gameId = command->getGameId();
     gameMonitor.startGame(playerId, gameId, sendQueue);
     std::cout << "[SG HANDLER] Game started" << std::endl;
     inGame = true;
